+++ config/dev.exs
@@ -6,4 +6,6 @@ config :sirko, :scheduler,
   # 1 hr
   remove_stale_data_every: 3600 * 1000
 
-config :sirko, :engine, max_pages_in_prediction: 2
+config :sirko, :engine,
+  max_pages_in_prediction: 4,
+  confidence_threshold: 0.1