+++ lib/sirko/entry.ex
@@ -3,6 +3,8 @@ defmodule Sirko.Entry do
   Describes data which gets sent by the client.
   """
 
+  @enforce_keys [:current_path, :referrer_path]
+
   defstruct current_path: nil,
             referrer_path: nil,
             assets: []
@@ -12,4 +14,9 @@ defmodule Sirko.Entry do
           referrer_path: String.t(),
           assets: [String.t()]
         }
+
+  @spec new(attrs :: Keyword.t()) :: t()
+  def new(attrs) do
+    struct(__MODULE__, attrs)
+  end
 end