+++ rel/commands/import_ga.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+release_ctl eval --mfa "Mix.Tasks.Sirko.ImportGa.run/1" --argv -- "$@"